refactor: abstract test cases over mcp client only instead over transport
